% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb2_get_url_to_query.R
\name{pxweb2_GET_url_to_px}
\alias{pxweb2_GET_url_to_px}
\title{Converts the GET URL from SCB to table_id and selected query into R}
\usage{
pxweb2_GET_url_to_px(url)
}
\arguments{
\item{url}{A SCB GET API URL. Located under Spara -> API-uttag -> GET}
}
\value{
A list with two elements:
\describe{
\item{table}{Table id.}
\item{query}{Query list suitable for pxweb_get_data().}
}
}
\description{
Extracts table_id and valuecodes from a Statistics Sweden API URL
and converts them to a list suitable for use with pxweb_get_data()
}
\examples{
x <- pxweb2_GET_url_to_px(url = "https://statistikdatabasen.scb.se/api/v2/tables/TAB5457/data?lang=sv&outputFormat=json-stat2&valuecodes[ContentsCode]=*&valuecodes[Tid]=2013,2014,2015,2016,2017,2018,2019,2020,2021,2022,2023,2024,2025&valuecodes[Region]=21&codelist[Region]=vs_A_L\%C3\%A4nRMI&heading=ContentsCode,Tid&stub=Region")

x$table
x$query

}
