% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb2r.R
\name{pxweb2_table_updated}
\alias{pxweb2_table_updated}
\title{Get when a table was last updated}
\usage{
pxweb2_table_updated(
  table,
  base_url = "https://statistikdatabasen.scb.se/api/v2/tables/"
)
}
\arguments{
\item{table}{Table id or metadata object.}

\item{base_url}{Base URL of the PxWeb API v2.}
}
\value{
Timestamp (ISO 8601) as a string, e.g. \code{"2026-06-02T00:59:31Z"}, or
\code{NA} if the value is missing.
}
\description{
Get when a table was last updated
}
