% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb2r-package.R
\docType{package}
\name{pxweb2r-package}
\alias{pxweb2r}
\alias{pxweb2r-package}
\title{pxweb2r: Get Data From a PxWeb API v2}
\description{
Tools for fetching data and metadata from a PxWeb API v2, by default Statistics Sweden's statistical database. Handles automatic chunking of large requests, label values in queries, wildcards, "latest period", aggregations, and splitting DeSO and RegSO by municipality. Results are returned as tidy tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FaluPeppe/pxweb2r}
  \item Report bugs at \url{https://github.com/FaluPeppe/pxweb2r/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

}
\keyword{internal}
