% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxweb2r-package.R
\docType{package}
\name{pxweb2r-package}
\alias{pxweb2r}
\alias{pxweb2r-package}
\title{pxweb2r: Hämta Data Från PxWeb API v2}
\description{
Verktyg för att hämta data och metadata från ett PxWeb API v2, som standard SCB:s statistikdatabas. Hanterar automatisk chunkning av stora uttag, klartextvärden i queries, wildcards, "senaste tid", aggregeringar samt uppdelning av DeSO och RegSO på kommun. Resultatet returneras som tidy tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FaluPeppe/pxweb2r}
  \item Report bugs at \url{https://github.com/FaluPeppe/pxweb2r/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

}
\keyword{internal}
